backend/requirements.txt
fastapi
uvicorn
python-multipart
spacy
pdfminer.six
docx2txt
scikit-learn
joblib
